final commit before demo  (#55)
* merge additional implementation into master (#49)

* added test entire controller and service for task, improve handle for mappings

* resolve mapping user in board and task, merged dev into my branch

* add password

* remove useless getter and setter

* implement findTaskByTaskCodeAndActiveTrue only in task service, start creating automatic code, handle control for the name in the taskController

* Add query getTaskNameAndIdConcatenated

* Changed password

* Implement create

* Remove useless properties

* Remove code in all class, now using id for the query in db. Fix all put and create method in the single relationship.

* Modify some message in controllers

* Handle find for delete logic, now return in DTO only daughter active

* Handle delete

* Change relationship task-lane-board, and handle exception for query for DB

* Implement logical delete in all classes

* change password

* Fix some mapper because merge dont work

* merged dev into change-relationship, fixed list user endpoint and now return only ROLE_USER

* change password

* after creation of board, there are a single lane TODO, task can't be associated in another lane than not belongs to board

* fix all find by repository

* add MaxChar rules in the controller and in database, and remove additional catch because the controllerAdvice handle all Controllers Exceptions

* implement new regex for the special character

* implements delete in cascade from user to relationship many to many

* change password

* add coding and decoding base64

* add JPA Projections, for improve performance for the query in task

* change password

---------


* merge additional-implementation into master (#52)

* added test entire controller and service for task, improve handle for mappings

* resolve mapping user in board and task, merged dev into my branch

* add password

* remove useless getter and setter

* implement findTaskByTaskCodeAndActiveTrue only in task service, start creating automatic code, handle control for the name in the taskController

* Add query getTaskNameAndIdConcatenated

* Changed password

* Implement create

* Remove useless properties

* Remove code in all class, now using id for the query in db. Fix all put and create method in the single relationship.

* Modify some message in controllers

* Handle find for delete logic, now return in DTO only daughter active

* Handle delete

* Change relationship task-lane-board, and handle exception for query for DB

* Implement logical delete in all classes

* change password

* Fix some mapper because merge dont work

* merged dev into change-relationship, fixed list user endpoint and now return only ROLE_USER

* change password

* after creation of board, there are a single lane TODO, task can't be associated in another lane than not belongs to board

* fix all find by repository

* add MaxChar rules in the controller and in database, and remove additional catch because the controllerAdvice handle all Controllers Exceptions

* implement new regex for the special character

* implements delete in cascade from user to relationship many to many

* change password

* add coding and decoding base64

* add JPA Projections, for improve performance for the query in task

* change password

* implement loggers

* adjust file xml

* add loggers in all classes

* change password

* adjust handle exceptions for the loggers

* change name class logger

* change nomenclature logger

* better explain

---------


* removed wrong return on boards service

* Implement create

* improve name variable

* add exception

* implement task user modify

* added checks for userboard update and delete methods

* implementations for update and delete works properly

* managed the security config files with access based on role on some endpoints

* implement task permissions

* add permission to Lane

* better format of the date and added a todo

* Fixed update, delete still has to be fixed

* Fixed missing permissions on delete and update

---------

+++ src/main/java/com/fincons/taskmanager/service/attachmentService/impl/AttachmentServiceImpl.java
@@ -1,17 +1,18 @@
 package com.fincons.taskmanager.service.attachmentService.impl;
 
 
-import com.fincons.taskmanager.entity.Attachment;
-import com.fincons.taskmanager.entity.AttachmentDownload;
-import com.fincons.taskmanager.entity.Task;
+import com.fincons.taskmanager.entity.*;
 import com.fincons.taskmanager.enums.FileExtension;
 import com.fincons.taskmanager.exception.ResourceNotFoundException;
+import com.fincons.taskmanager.exception.RoleException;
 import com.fincons.taskmanager.repository.AttachmentRepository;
 import com.fincons.taskmanager.service.attachmentService.AttachmentService;
 import com.fincons.taskmanager.service.taskService.impl.TaskServiceImpl;
+import com.fincons.taskmanager.service.userBoardService.UserBoardService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -22,6 +23,10 @@ import java.util.Objects;
 @Service
 public class AttachmentServiceImpl implements AttachmentService {
 
+
+    private static final String EDITOR = "EDITOR";
+    @Autowired
+    private UserBoardService userBoardService;
     @Autowired
     private AttachmentRepository attachmentRepository;
     @Autowired
@@ -39,8 +44,27 @@ public class AttachmentServiceImpl implements AttachmentService {
                 AttachmentDecoding.decodeToString(attachmentExisting), attachmentExisting);
     }
     @Override
-    public Attachment uploadAttachment(Long taskId, MultipartFile file) throws IOException {
+    public Attachment uploadAttachment(Long taskId, MultipartFile file) throws IOException, RoleException {
         Task task = taskServiceImpl.validateTaskById(taskId);
+
+        String loggedUser = SecurityContextHolder.getContext().getAuthentication().getName();
+        boolean userAssociated = task.getTasksUsers().stream()
+                .anyMatch(taskUser -> Objects.equals(taskUser.getUser().getEmail(), loggedUser));
+        Lane laneToGetBoard = task.getLane();
+        long boardId = laneToGetBoard.getBoard().getBoardId();
+        List<UserBoard> userBoards = userBoardService.findBoardsByUser(loggedUser);
+        List<UserBoard> userInBoard = userBoards.stream().filter(userBoardToCheck -> userBoardToCheck.getBoard().getBoardId() == boardId).toList();
+
+        UserBoard userBoard = userInBoard.get(0);
+        boolean userIsEditor = false;
+        String roleCode = userBoard.getRoleCode();
+        log.info(userBoard.getRoleCode());
+        if(EDITOR.equals(roleCode)){
+            userIsEditor = true;
+        }
+        if (!userAssociated && !userIsEditor){
+            throw new RoleException("You are not associated with this task and cannot delete it.");
+        }
         String attachmentName = getFileName(file);
         String extension = getFileExtension(file);
         compatibleExtension(extension);
@@ -51,8 +75,27 @@ public class AttachmentServiceImpl implements AttachmentService {
         return attachment;
     }
     @Override
-    public void deleteAttachmentById(Long attachmentId) {
+    public void deleteAttachmentById(Long attachmentId) throws RoleException {
         Attachment attachment = validateAttachmentById(attachmentId);
+        Task task = taskServiceImpl.validateTaskById(attachment.getTask().getTaskId());
+        String loggedUser = SecurityContextHolder.getContext().getAuthentication().getName();
+        boolean userAssociated = task.getTasksUsers().stream()
+                .anyMatch(taskUser -> Objects.equals(taskUser.getUser().getEmail(), loggedUser));
+        Lane laneToGetBoard = task.getLane();
+        long boardId = laneToGetBoard.getBoard().getBoardId();
+        List<UserBoard> userBoards = userBoardService.findBoardsByUser(loggedUser);
+        List<UserBoard> userInBoard = userBoards.stream().filter(userBoardToCheck -> userBoardToCheck.getBoard().getBoardId() == boardId).toList();
+
+        UserBoard userBoard = userInBoard.get(0);
+        boolean userIsEditor = false;
+        String roleCode = userBoard.getRoleCode();
+        log.info(userBoard.getRoleCode());
+        if(EDITOR.equals(roleCode)){
+            userIsEditor = true;
+        }
+        if (!userAssociated && !userIsEditor){
+            throw new RoleException("You are not associated with this task and cannot delete attachments on it.");
+        }
         attachmentRepository.delete(attachment);
         log.info("Attachment with ID {} deleted from the repository.", attachment.getAttachmentId());
     }
