+++ src/main/java/com/fincons/taskmanager/service/userBoardService/impl/UserBoardServiceImpl.java
@@ -3,12 +3,17 @@ package com.fincons.taskmanager.service.userBoardService.impl;
 import com.fincons.taskmanager.entity.*;
 import com.fincons.taskmanager.exception.DuplicateException;
 import com.fincons.taskmanager.exception.ResourceNotFoundException;
+import com.fincons.taskmanager.exception.RoleException;
 import com.fincons.taskmanager.repository.BoardRepository;
 import com.fincons.taskmanager.repository.RoleRepository;
 import com.fincons.taskmanager.repository.UserBoardRepository;
 import com.fincons.taskmanager.repository.UserRepository;
 import com.fincons.taskmanager.service.userBoardService.UserBoardService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,6 +22,10 @@ import java.util.Objects;
 @Service
 public class UserBoardServiceImpl implements UserBoardService {
 
+    private static final Logger log = LogManager.getLogger(UserBoardServiceImpl.class);
+    private static final String EDITOR = "EDITOR";
+    private static final String ROLEADMIN = "ROLE_ADMIN";
+    private static final String ROLEUSER = "ROLE_USER";
     @Autowired
     private UserBoardRepository userBoardRepository;
     @Autowired
@@ -33,20 +42,48 @@ public class UserBoardServiceImpl implements UserBoardService {
     }
 
     @Override
-    public UserBoard createUserBoard(UserBoard userBoard) {
+    public UserBoard createUserBoard(UserBoard userBoard) throws RoleException {
 
         User existingUser = validateUserByEmail(userBoard.getUser().getEmail());
         Board existingBoard = validateBoardById(userBoard.getBoard().getBoardId());
-
         checkDuplicateUserBoardExist(existingUser, existingBoard);
-        UserBoard newUserBoard = new UserBoard(existingUser, existingBoard, userBoard.getRoleCode());
-        userBoardRepository.save(newUserBoard);
 
-        return newUserBoard;
+        GrantedAuthority authority = SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+                .findFirst()
+                .orElse(null);
+        String roleAsString = (authority != null) ? authority.getAuthority() : "is null";
+        boolean canUserBoardBeCreated = false;
+
+        if (Objects.equals(roleAsString, ROLEADMIN)) {
+            canUserBoardBeCreated = true;
+        }
+
+        if (Objects.equals(roleAsString, ROLEUSER)) {
+            String loggedUser = SecurityContextHolder.getContext().getAuthentication().getName();
+            List<User> userIsEditor = existingBoard.getUsersBoards().stream()
+                    .filter(userBoardToCreate -> EDITOR.equals(userBoardToCreate.getRoleCode()))
+                    .map(UserBoard::getUser)
+                    .toList();
+            boolean isEditorEmailValidated = userIsEditor.stream()
+                    .anyMatch(user -> Objects.equals(user.getEmail(), loggedUser));
+            if (isEditorEmailValidated) {
+                canUserBoardBeCreated = true;
+            } else {
+                throw new RoleException("You need to be an EDITOR to create a relationship");
+            }
+        }
+        if (canUserBoardBeCreated) {
+            UserBoard newUserBoard = new UserBoard(existingUser, existingBoard, userBoard.getRoleCode());
+            userBoardRepository.save(newUserBoard);
+            return newUserBoard;
+        } else {
+            log.error("UNEXPECTED ERROR");
+            throw new IllegalArgumentException("UNEXPECTED ERROR");
+        }
     }
 
     @Override
-    public UserBoard updateUserBoard(String email, Long boardId, UserBoard userBoard) {
+    public UserBoard updateUserBoard(String email, Long boardId, UserBoard userBoard) throws RoleException {
 
         User existingUser = validateUserByEmail(email);
         Board existingBoard = validateBoardById(boardId);
@@ -54,24 +91,81 @@ public class UserBoardServiceImpl implements UserBoardService {
 
         User userToUpdate = validateUserByEmail(userBoard.getUser().getEmail());
         Board boardToUpdate = validateBoardById(userBoard.getBoard().getBoardId());
-        validateUserBoardNotExistRelationship(userToUpdate, boardToUpdate);
 
-        userBoardExist.setUser(userToUpdate);
-        userBoardExist.setBoard(boardToUpdate);
-        userBoardExist.setRoleCode(userBoard.getRoleCode());
-        userBoardRepository.save(userBoardExist);
+        GrantedAuthority authority = SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+                .findFirst()
+                .orElse(null);
+        String roleAsString = (authority != null) ? authority.getAuthority() : "is null";
+        boolean canUserBoardBeUpdated = false;
 
-        return userBoardExist;
+        if (Objects.equals(roleAsString, ROLEADMIN)) {
+            canUserBoardBeUpdated = true;
+        }
 
+        if (Objects.equals(roleAsString, ROLEUSER)) {
+            String loggedUser = SecurityContextHolder.getContext().getAuthentication().getName();
+            List<User> userIsEditor = boardToUpdate.getUsersBoards().stream()
+                    .filter(userBoardToFilter -> EDITOR.equals(userBoardToFilter.getRoleCode()))
+                    .map(UserBoard::getUser)
+                    .toList();
+            boolean isEditorEmailValidated = userIsEditor.stream()
+                    .anyMatch(user -> Objects.equals(user.getEmail(), loggedUser));
+            if (isEditorEmailValidated) {
+                canUserBoardBeUpdated = true;
+            } else {
+                throw new RoleException("You need to be an EDITOR to update a relationship");
+            }
+        }
+        if (canUserBoardBeUpdated) {
+            userBoardExist.setUser(userToUpdate);
+            userBoardExist.setBoard(boardToUpdate);
+            userBoardExist.setRoleCode(userBoard.getRoleCode());
+            return userBoardRepository.save(userBoardExist);
+        } else {
+            log.error("UNEXPECTED ERROR");
+            throw new IllegalArgumentException("UNEXPECTED ERROR");
+        }
     }
+
     @Override
-    public UserBoard deleteUserBoard(String email, Long boardId) {
+    public UserBoard deleteUserBoard(String email, Long boardId) throws RoleException {
         User existingUser = validateUserByEmail(email);
         Board existingBoard = validateBoardById(boardId);
         UserBoard userBoardExist = validateUserBoardRelationship(existingUser, existingBoard);
-        userBoardRepository.delete(userBoardExist);
-        return userBoardExist;
+
+        GrantedAuthority authority = SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+                .findFirst()
+                .orElse(null);
+        String roleAsString = (authority != null) ? authority.getAuthority() : "is null";
+        boolean canUserBeDeleted = false;
+
+        if (Objects.equals(roleAsString, ROLEADMIN)) {
+            canUserBeDeleted = true;
+        }
+
+        if (Objects.equals(roleAsString, ROLEUSER)) {
+            String loggedUser = SecurityContextHolder.getContext().getAuthentication().getName();
+            List<User> userIsEditor = existingBoard.getUsersBoards().stream()
+                    .filter(userBoardToFilter -> EDITOR.equals(userBoardToFilter.getRoleCode()))
+                    .map(UserBoard::getUser)
+                    .toList();
+            boolean isEditorEmailValidated = userIsEditor.stream()
+                    .anyMatch(user -> Objects.equals(user.getEmail(), loggedUser));
+            if (isEditorEmailValidated) {
+                canUserBeDeleted = true;
+            } else {
+                throw new RoleException("You need to be an EDITOR to delete a relationship");
+            }
+        }
+        if (canUserBeDeleted) {
+            userBoardRepository.delete(userBoardExist);
+            return userBoardExist;
+        } else {
+            log.error("UNEXPECTED ERROR");
+            throw new IllegalArgumentException("UNEXPECTED ERROR");
+        }
     }
+
     private User validateUserByEmail(String email) {
         User existingUser = userRepository.findByEmail(email);
         if (Objects.isNull(existingUser)) {
@@ -79,6 +173,7 @@ public class UserBoardServiceImpl implements UserBoardService {
         }
         return existingUser;
     }
+
     private Board validateBoardById(Long id) {
         Board existingBoard = boardRepository.findBoardByBoardIdAndActiveTrue(id);
         if (Objects.isNull(existingBoard)) {
@@ -86,6 +181,7 @@ public class UserBoardServiceImpl implements UserBoardService {
         }
         return existingBoard;
     }
+
     private void checkDuplicateUserBoardExist(User user, Board board) {
 
         boolean userBoardExist = userBoardRepository.existsByUserAndBoard(user, board);
@@ -95,6 +191,7 @@ public class UserBoardServiceImpl implements UserBoardService {
                     "USER: " + user.getEmail() + " and BOARD: " + board.getBoardId());
         }
     }
+
     private UserBoard validateUserBoardRelationship(User user, Board board) {
         UserBoard userBoardExist = userBoardRepository.findByUserEmailAndBoardBoardId(user.getEmail(), board.getBoardId());
         if (Objects.isNull(userBoardExist)) {
@@ -103,6 +200,7 @@ public class UserBoardServiceImpl implements UserBoardService {
         }
         return userBoardExist;
     }
+
     private void validateUserBoardNotExistRelationship(User user, Board board) {
         UserBoard userBoardExist = userBoardRepository.findByUserEmailAndBoardBoardId(user.getEmail(), board.getBoardId());
         if (!Objects.isNull(userBoardExist)) {
